+++ apps/vast/CHANGELOG.md
@@ -19,6 +19,17 @@ Every entry has a category for which we use the following visual abbreviations:
   fields are separated by commas.
   [#136](https://github.com/tenzir/threatbus/pull/136)
 
+- ⚠️ `pyvast-threatbus` now uses
+  [Dynaconf](https://github.com/rochacbruno/dynaconf) for configuration
+  management. Configuration via a config file works exactly as it has worked
+  before. Users can provide a path to the config file using the `-c` option.
+  `pyvast-threatbus` now considers files named `config.yaml` and `config.yml` as
+  default configs if located in the same directory. Additionally,
+  `pyvast-threatbus` now supports configration via environment variables and
+  `.dotenv`. Env vars need to be prefixed with `PYVAST_THREATBUS_` to be
+  respected and always take precedence over values in config files.
+  [#133](https://github.com/tenzir/threatbus/pull/133)
+
 ## [2021.06.24]
 
 - ⚡️ `pyvast-threatbus` now uses point queries over substring queries for
@@ -26,6 +37,7 @@ Every entry has a category for which we use the following visual abbreviations:
   in less matches than before. E.g., a URI indicator `tenzir.com` that
   used to match `docs.tenzir.com` as well as `https://tenzir.com` now
   only matches exactly the indicator.
+  [#130](https://github.com/tenzir/threatbus/pull/130)
 
 - 🎁 `pyvast-threatbus` now collects metrics about received indicators that are
   about to be matched retrospectively against VAST. The new metric is called
@@ -33,7 +45,7 @@ Every entry has a category for which we use the following visual abbreviations:
   building up. A backlog builds when `pyvast-threatbus` hits the
   user-configured limit of max backgroud tasks while at the same time VAST
   responds slowly to the issued queries.
-  [#129]https://github.com/tenzir/threatbus/pull/129
+  [#129](https://github.com/tenzir/threatbus/pull/129)
 
 - 🎁 `pyvast-threatbus` now comes with its own Dockerfile. Pre-built images are
   available on [Dockerhub](https://hub.docker.com/r/tenzir/pyvast-threatbus).